Split background chooser to own component
